REFACTOR: Remove shared code from Rest controllers
