+++ inc/Modules/Core/Assets.php
@@ -10,7 +10,7 @@ declare( strict_types = 1 );
 namespace OneMedia\Modules\Core;
 
 use OneMedia\Contracts\Interfaces\Registrable;
-use OneMedia\Modules\Rest\Media_Sharing_Controller;
+use OneMedia\Modules\Rest\Utils as Rest_Utils;
 use OneMedia\Modules\Settings\Settings;
 
 /**
@@ -75,13 +75,13 @@ final class Assets implements Registrable {
 				'siteName'         => get_bloginfo( 'name' ),
 				'ajaxUrl'          => admin_url( 'admin-ajax.php' ),
 				'uploadNonce'      => wp_create_nonce( 'onemedia_upload_media' ),
-				'allowedMimeTypes' => Media_Sharing_Controller::get_supported_mime_types(),
+				'allowedMimeTypes' => Rest_Utils::get_supported_mime_types(),
 				'mediaSyncNonce'   => wp_create_nonce( 'onemedia_check_sync_status' ),
 				'allLabel'         => __( 'All media', 'onemedia' ),
 				'syncLabel'        => __( 'Synced', 'onemedia' ),
 				'notSyncLabel'     => __( 'Not Synced', 'onemedia' ),
 				'filterLabel'      => __( 'Sync Status', 'onemedia' ),
-				'syncStatus'       => Media_Sharing_Controller::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY,
+				'syncStatus'       => Rest_Utils::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY,
 			];
 		}
