+++ inc/Modules/Rest/Media_Sharing_Controller.php
@@ -9,6 +9,7 @@ namespace OneMedia\Modules\Rest;
 
 use OneMedia\Modules\MediaSharing\MediaActions;
 use OneMedia\Modules\MediaSharing\MediaReplacement;
+use OneMedia\Modules\Rest\Utils as Rest_Utils;
 use OneMedia\Modules\Settings\Settings;
 use OneMedia\Modules\Taxonomies\Term_Restriction;
 use WP_REST_Server;
@@ -32,26 +33,6 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 	 */
 	public const FETCH_MEDIA_REQUEST_TIMEOUT = 30;
 
-	/**
-	 * OneMedia sync status postmeta key.
-	 *
-	 * @var string
-	 */
-	public const ONEMEDIA_SYNC_STATUS_POSTMETA_KEY = 'onemedia_sync_status';
-
-	/**
-	 * OneMedia sync sites postmeta key.
-	 *
-	 * @var string
-	 */
-	public const ONEMEDIA_SYNC_SITES_POSTMETA_KEY = 'onemedia_sync_sites';
-
-	/**
-	 * Is OneMedia sync postmeta key.
-	 *
-	 * @var string
-	 */
-	public const IS_ONEMEDIA_SYNC_POSTMETA_KEY = 'is_onemedia_sync';
 
 	/**
 	 * OneMedia brand site to governing site attachment key map.
@@ -59,32 +40,6 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 	 * @var string
 	 */
 	public const ONEMEDIA_ATTACHMENT_KEY_MAP_OPTION = 'onemedia_attachment_key_map';
-
-	/**
-	 * Brand sites synced media option.
-	 *
-	 * @var string
-	 */
-	public const BRAND_SITES_SYNCED_MEDIA_OPTION = 'onemedia_brand_sites_synced_media';
-
-
-	/**
-	 * Allowed mime types array.
-	 *
-	 * This is a list of potentially supported mime types, any unsupported mime types will
-	 * be removed during usage, on that particular server.
-	 *
-	 * @var array
-	 */
-	public const ALLOWED_MIME_TYPES = [
-		'image/jpg',
-		'image/jpeg',
-		'image/png',
-		'image/gif',
-		'image/webp',
-		'image/bmp',
-		'image/svg+xml',
-	];
 
 	/**
 	 * Register REST API routes.
@@ -348,7 +303,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 	 */
 	public function get_brand_sites_synced_media(): \WP_Error|\WP_REST_Response {
 		// Get the sync option data.
-		$brand_sites_synced_media = self::fetch_brand_sites_synced_media();
+		$brand_sites_synced_media = Rest_Utils::fetch_brand_sites_synced_media();
 
 		if ( ! is_array( $brand_sites_synced_media ) ) {
 			$brand_sites_synced_media = [];
@@ -410,9 +365,9 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 		}
 
 		// Delete the metadata for the attachment.
-		delete_post_meta( $attachment_id, self::IS_ONEMEDIA_SYNC_POSTMETA_KEY );
-		delete_post_meta( $attachment_id, self::ONEMEDIA_SYNC_SITES_POSTMETA_KEY );
-		delete_post_meta( $attachment_id, self::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY );
+		delete_post_meta( $attachment_id, Rest_Utils::IS_ONEMEDIA_SYNC_POSTMETA_KEY );
+		delete_post_meta( $attachment_id, Rest_Utils::ONEMEDIA_SYNC_SITES_POSTMETA_KEY );
+		delete_post_meta( $attachment_id, Rest_Utils::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY );
 
 		return rest_ensure_response(
 			[
@@ -499,7 +454,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 		];
 
 		// Decode HTML entities in title.
-		$attachment_post['post_title'] = self::decode_filename( $attachment_post['post_title'] );
+		$attachment_post['post_title'] = Rest_Utils::decode_filename( $attachment_post['post_title'] );
 
 		// Download the file from the URL and save it to the uploads directory and replace the old file.
 		$uploads  = wp_get_upload_dir();
@@ -633,7 +588,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			'orderby'        => 'date',
 			'order'          => 'DESC',
 			'menu_order'     => 'ASC',
-			'post_mime_type' => self::get_supported_mime_types(),
+			'post_mime_type' => Rest_Utils::get_supported_mime_types(),
 		];
 
 		// Add search functionality.
@@ -671,7 +626,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			while ( $query->have_posts() ) {
 				$query->the_post();
 				$post_id       = get_the_ID();
-				$title         = self::decode_filename( get_the_title( $post_id ) );
+				$title         = Rest_Utils::decode_filename( get_the_title( $post_id ) );
 				$media_files[] = [
 					'id'        => $post_id,
 					'url'       => wp_get_attachment_url( $post_id ),
@@ -791,7 +746,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			}
 
 			// Mime type should be one of the supported mime types.
-			if ( ! in_array( $media['mime_type'], self::get_supported_mime_types(), true ) ) {
+			if ( ! in_array( $media['mime_type'], Rest_Utils::get_supported_mime_types(), true ) ) {
 				return new \WP_Error(
 					'invalid_mime_type',
 					__( 'Invalid mime type provided.', 'onemedia' ),
@@ -804,10 +759,10 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 
 			// For each media file its sync is checked then add meta data is_onemedia_sync to be true and onemedia_sync_sites meta to array of sites where it is synced.
 			if ( 'sync' === $sync_option ) {
-				update_post_meta( $media['id'], self::IS_ONEMEDIA_SYNC_POSTMETA_KEY, true );
+				update_post_meta( $media['id'], Rest_Utils::IS_ONEMEDIA_SYNC_POSTMETA_KEY, true );
 			} else {
 				// If not syncing, set the sync status to false.
-				update_post_meta( $media['id'], self::IS_ONEMEDIA_SYNC_POSTMETA_KEY, 0 );
+				update_post_meta( $media['id'], Rest_Utils::IS_ONEMEDIA_SYNC_POSTMETA_KEY, 0 );
 			}
 
 			// Share the attachment metadata with the brand sites.
@@ -855,7 +810,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 				continue;
 			}
 
-			$site_name = Basic_Options_Controller::get_sitename_by_url( $site_url );
+			$site_name = Settings::get_sitename_by_url( $site_url );
 
 			// Find the site in all brand sites to get its API token.
 			foreach ( $all_brand_sites as $site_data ) {
@@ -926,7 +881,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			foreach ( $media_response_list as $media ) {
 				// Get onemedia_sync_sites meta for each parent_id media.
 				$parent_id           = $media['parent_id'];
-				$onemedia_sync_sites = Basic_Options_Controller::get_sync_sites_postmeta( $parent_id );
+				$onemedia_sync_sites = Rest_Utils::health_check_attachment_brand_sites( $parent_id );
 
 				if ( ! is_array( $onemedia_sync_sites ) ) {
 					$onemedia_sync_sites = [];
@@ -939,14 +894,14 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 				];
 
 				// Update onemedia_sync_sites meta for each parent_id media.
-				update_post_meta( $parent_id, self::ONEMEDIA_SYNC_SITES_POSTMETA_KEY, $onemedia_sync_sites );
+				update_post_meta( $parent_id, Rest_Utils::ONEMEDIA_SYNC_SITES_POSTMETA_KEY, $onemedia_sync_sites );
 
 				// Create option to store siteurl, parent media id and brand site media id.
 				if ( 'sync' !== $sync_option ) {
 					continue;
 				}
 
-				$brand_sites_synced_media = self::fetch_brand_sites_synced_media();
+				$brand_sites_synced_media = Rest_Utils::fetch_brand_sites_synced_media();
 
 				if ( ! is_array( $brand_sites_synced_media ) ) {
 					$brand_sites_synced_media = [];
@@ -967,12 +922,12 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 				);
 
 				// Update the synced media mapping option only if there is a change.
-				$saved_brand_sites_synced_media = self::fetch_brand_sites_synced_media();
+				$saved_brand_sites_synced_media = Rest_Utils::fetch_brand_sites_synced_media();
 				if ( hash_equals( wp_json_encode( $saved_brand_sites_synced_media ), wp_json_encode( $brand_sites_synced_media ) ) ) {
 					continue;
 				}
 
-				$success = update_option( self::BRAND_SITES_SYNCED_MEDIA_OPTION, $brand_sites_synced_media );
+				$success = update_option( Rest_Utils::BRAND_SITES_SYNCED_MEDIA_OPTION, $brand_sites_synced_media );
 
 				if ( $success ) {
 					continue;
@@ -1007,7 +962,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 				'status'               => 200,
 				'success'              => true,
 				'success_response'     => $success_response,
-				'onemedia_sync_option' => self::fetch_brand_sites_synced_media(),
+				'onemedia_sync_option' => Rest_Utils::fetch_brand_sites_synced_media(),
 			]
 		);
 	}
@@ -1101,7 +1056,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			];
 
 			// Decode HTML entities in title.
-			$attachment_metadata['post_title'] = self::decode_filename( $attachment_metadata['post_title'] );
+			$attachment_metadata['post_title'] = Rest_Utils::decode_filename( $attachment_metadata['post_title'] );
 
 			// Get governing to brand site attachment mapping to prevent duplicate media files.
 			$attachment_key_map = self::get_attachment_key_map();
@@ -1119,7 +1074,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 					if ( 'sync' === $sync_status ) {
 						// Add attachment metadata for sync status.
 						if ( $sync_status ) {
-							update_post_meta( $saved_attachment_id, self::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
+							update_post_meta( $saved_attachment_id, Rest_Utils::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
 						}
 
 						// Update the existing attachment with new metadata if any changes are present.
@@ -1133,7 +1088,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 						'url'          => wp_get_attachment_url( $saved_attachment_id ),
 						'title'        => $media_title,
 						'parent_id'    => $parent_media_id,
-						'parent_terms' => self::get_onemedia_attachment_terms( $saved_attachment_id ) ?? [],
+						'parent_terms' => Rest_Utils::get_onemedia_attachment_terms( $saved_attachment_id ) ?? [],
 					];
 				} else {
 					// Media already shared but in different configuration. Convert media from non-sync to sync.
@@ -1157,7 +1112,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 
 					// Add attachment metadata for sync status.
 					if ( $sync_status ) {
-						update_post_meta( $attachment_id, self::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
+						update_post_meta( $attachment_id, Rest_Utils::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
 					}
 
 					// Update the existing attachment with new metadata if any changes are present.
@@ -1168,13 +1123,13 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 						'url'          => wp_get_attachment_url( $attachment_id ),
 						'title'        => $media_title,
 						'parent_id'    => $parent_media_id,
-						'parent_terms' => self::get_onemedia_attachment_terms( $attachment_id ) ?? [],
+						'parent_terms' => Rest_Utils::get_onemedia_attachment_terms( $attachment_id ) ?? [],
 					];
 				}
 			} else { // New media file, not shared previously.
 
 				// Check for unsupported mime types before adding.
-				if ( ! in_array( $media_mime_type, self::get_supported_mime_types(), true ) ) {
+				if ( ! in_array( $media_mime_type, Rest_Utils::get_supported_mime_types(), true ) ) {
 					$unsupported_file_types[] = self::get_file_type_label( $media_mime_type );
 					continue;
 				}
@@ -1254,7 +1209,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 					'url'          => wp_get_attachment_url( $attachment_id ),
 					'title'        => $media_title,
 					'parent_id'    => $parent_media_id,
-					'parent_terms' => self::get_onemedia_attachment_terms( $attachment_id ) ?? [],
+					'parent_terms' => Rest_Utils::get_onemedia_attachment_terms( $attachment_id ) ?? [],
 				];
 			}
 		}
@@ -1270,7 +1225,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 				sprintf(
 					/* translators: %1$s: Site name, %2$s: Unsupported file types. */
 					__( '%1$s site doesn\'t support the following file type(s): %2$s.', 'onemedia' ),
-					Basic_Options_Controller::get_sitename_by_url( get_bloginfo( 'url' ) ),
+					Settings::get_sitename_by_url( get_bloginfo( 'url' ) ),
 					$unsupported_file_types,
 				),
 				[
@@ -1348,7 +1303,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 		$attachment_file_type = get_post_mime_type( $attachment_id );
 
 		// Validate file type.
-		if ( ! in_array( $attachment_file_type, self::get_supported_mime_types(), true ) ) {
+		if ( ! in_array( $attachment_file_type, Rest_Utils::get_supported_mime_types(), true ) ) {
 			return new \WP_Error(
 				'invalid_data',
 				__( 'Invalid attachment file type provided.', 'onemedia' ),
@@ -1389,11 +1344,11 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			}
 
 			// Assign the meta key to the attachment.
-			update_post_meta( $attachment_id, self::IS_ONEMEDIA_SYNC_POSTMETA_KEY, true );
+			update_post_meta( $attachment_id, Rest_Utils::IS_ONEMEDIA_SYNC_POSTMETA_KEY, true );
 
 			// Convert non sync to sync media if it was previously shared as non sync.
 			// Check if the media is already shared in non-sync mode.
-			$shared_sites = Basic_Options_Controller::get_sync_sites_postmeta( $attachment_id );
+			$shared_sites = Rest_Utils::health_check_attachment_brand_sites( $attachment_id );
 			if ( is_array( $shared_sites ) ) {
 				// Perform the media sync operation here.
 				$brand_site_prefix = '/wp-json/' . Abstract_REST_Controller::NAMESPACE . '/add-media';
@@ -1412,7 +1367,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 
 					// Strip the trailing slash.
 					$site_url   = rtrim( $site_url, '/' );
-					$site_name  = Basic_Options_Controller::get_sitename_by_url( $site_url );
+					$site_name  = Settings::get_sitename_by_url( $site_url );
 					$site_token = '';
 
 					if ( empty( $site_url ) ) {
@@ -1534,7 +1489,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 						$parent_id = $media['parent_id'];
 
 						// Create option to store siteurl, parent media id and brand site media id.
-						$brand_sites_synced_media = self::fetch_brand_sites_synced_media();
+						$brand_sites_synced_media = Rest_Utils::fetch_brand_sites_synced_media();
 						if ( ! is_array( $brand_sites_synced_media ) ) {
 							$brand_sites_synced_media = [];
 						}
@@ -1554,12 +1509,12 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 						);
 
 						// Update the synced media mapping option only if there is a change.
-						$saved_brand_sites_synced_media = self::fetch_brand_sites_synced_media();
+						$saved_brand_sites_synced_media = Rest_Utils::fetch_brand_sites_synced_media();
 						if ( hash_equals( wp_json_encode( $saved_brand_sites_synced_media ), wp_json_encode( $brand_sites_synced_media ) ) ) {
 							continue;
 						}
 
-						$success = update_option( self::BRAND_SITES_SYNCED_MEDIA_OPTION, $brand_sites_synced_media );
+						$success = update_option( Rest_Utils::BRAND_SITES_SYNCED_MEDIA_OPTION, $brand_sites_synced_media );
 
 						if ( ! $success ) {
 							$failed_sites[] = [
@@ -1594,14 +1549,14 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 						'status'               => 200,
 						'success_response'     => $success_response,
 						'attachment_id'        => $attachment_id,
-						'onemedia_sync_option' => self::fetch_brand_sites_synced_media(),
+						'onemedia_sync_option' => Rest_Utils::fetch_brand_sites_synced_media(),
 						'success'              => true,
 					]
 				);
 			}
 		} else {
 			// If not syncing, set the sync status to false.
-			update_post_meta( $attachment_id, self::IS_ONEMEDIA_SYNC_POSTMETA_KEY, 0 );
+			update_post_meta( $attachment_id, Rest_Utils::IS_ONEMEDIA_SYNC_POSTMETA_KEY, 0 );
 		}
 
 		// Return success response with attachment ID.
@@ -1891,7 +1846,7 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 
 		// Add attachment metadata for sync status.
 		if ( $sync_status ) {
-			update_post_meta( $attachment_id, self::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
+			update_post_meta( $attachment_id, Rest_Utils::ONEMEDIA_SYNC_STATUS_POSTMETA_KEY, $sync_status );
 		}
 
 		if ( ! function_exists( 'wp_generate_attachment_metadata' ) ) {
@@ -1989,17 +1944,6 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 	}
 
 	/**
-	 * Get brand site's synced media array.
-	 *
-	 * The structure of this array is different on governing and brand sites.
-	 *
-	 * @return array Array of brand site's synced media.
-	 */
-	public static function fetch_brand_sites_synced_media(): array {
-		return get_option( self::BRAND_SITES_SYNCED_MEDIA_OPTION, [] );
-	}
-
-	/**
 	 * Get attachment key map.
 	 *
 	 * This option contains the governing site to brand site attachment key map.
@@ -2012,17 +1956,6 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 			return [];
 		}
 		return get_option( self::ONEMEDIA_ATTACHMENT_KEY_MAP_OPTION, [] );
-	}
-
-	/**
-	 * Decode filename to handle special characters.
-	 *
-	 * @param string $filename The filename to decode.
-	 *
-	 * @return string The decoded filename.
-	 */
-	public static function decode_filename( string $filename ): string {
-		return html_entity_decode( $filename, ENT_QUOTES, 'UTF-8' );
 	}
 
 	/**
@@ -2046,21 +1979,6 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 	}
 
 	/**
-	 * Get supported mime types.
-	 *
-	 * @return array Array of supported mime types by the server.
-	 */
-	public static function get_supported_mime_types(): array {
-		$allowed_types = self::ALLOWED_MIME_TYPES;
-
-		// Remove any types that are not supported by the server.
-		$supported_types = array_values( get_allowed_mime_types() );
-		$allowed_types   = array_intersect( $allowed_types, $supported_types );
-
-		return $allowed_types;
-	}
-
-	/**
 	 * Check if a URL is valid or not.
 	 *
 	 * @param string $url The URL to check.
@@ -2077,24 +1995,5 @@ class Media_Sharing_Controller extends Abstract_REST_Controller {
 		$pattern = "/^https?:\\/\\/(?:www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b(?:[-a-zA-Z0-9()@:%_\\+.~#?&\\/=]*)$/";
 
 		return (bool) preg_match( $pattern, $url );
-	}
-
-	/**
-	 * Get OneMedia attachment terms.
-	 *
-	 * @param int|\WP_Post $attachment_id The attachment ID.
-	 *
-	 * @return array Array of terms.
-	 */
-	public static function get_onemedia_attachment_terms( int|\WP_Post $attachment_id ): array {
-		if ( ! $attachment_id ) {
-			return [];
-		}
-
-		$terms = get_the_terms( $attachment_id, Term_Restriction::ONEMEDIA_PLUGIN_TAXONOMY );
-		if ( is_wp_error( $terms ) || ! $terms ) {
-			return [];
-		}
-		return $terms;
 	}
 }
